Refactored existring intefaces and added new IDataProvider

+++ Lab 1 test/Interfaces/IEmployeeSalaryInfo.cs
@@ -5,6 +5,6 @@ namespace Lab_1_test.Interfaces
     {
         string FirstName { get; set; }
         string LastName { get; set; }
-        double Salary { get; set; }
+        decimal Salary { get; set; }
     }
 }
